refactor(status-bar): extract GitHub popover code into GitHubPopover.swift
Lifts ~1400 lines of file-private types out of StatusBarView.swift:
GitHubPopoverKind, GitHubStatusSegmentView, GitHubPopoverLayout,
PRDisplayStatus, GitHubPopoverRowModel, ConventionalCommitParser,
GitHubHoverPopoverViewController, GitHubPopoverContentView,
GitHubPopoverRowView, GitHubPopoverPreviewView,
GitHubPopoverFooterKeymapView.

Self-contained — no popover type references StatusBarView state.
Top-level types lose their private modifier (now internal) so
StatusBarView's cross-file references resolve. Nested privates
preserved.

StatusBarView.swift: 2408 → 990 lines (-58.9%).
